+++ src/test/java/com/api/request/model/CustomerAddress.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.request.model;
 
 public record CustomerAddress(
 	String flat_number,
