+++ src/test/java/com/api/tests/LoginAPITest.java
@@ -8,7 +8,7 @@ import java.io.IOException;
 
 import org.testng.annotations.Test;
 
-import com.api.pojo.UserCredentials;
+import com.api.request.model.UserCredentials;
 import com.api.utils.ConfigManager;
 import com.api.utils.SpecUtil;
