+++ src/test/java/com/api/utils/AuthTokenProvider.java
@@ -9,7 +9,7 @@ import static org.hamcrest.Matchers.*;
 import static com.api.constant.Role.*;
 
 import com.api.constant.Role;
-import com.api.pojo.UserCredentials;
+import com.api.request.model.UserCredentials;
 
 import io.restassured.http.ContentType;
